fix: remove internal `Transforms` namespace
Replace all `Transforms.X(editor, ...)` calls with `editor.X(...)` across
slate-react, slate-dom, and core internals. Move the operation engine
(`Transforms.transform`) to `core/transform-op.ts`. Inline all type
signatures in `BaseEditor` that previously derived from `Transforms`.
Delete the `Transforms` const, delegation layer, and `GeneralTransforms`.
