+++ packages/editor/src/slate-react/components/editable.tsx
@@ -18,7 +18,6 @@ import {
   Path,
   Range,
   Text,
-  Transforms,
   type DecoratedRange,
   type LeafPosition,
   type NodeEntry,
@@ -257,7 +256,7 @@ export const Editable = forwardRef(
             if (active) {
               document.execCommand('indent')
             } else {
-              Transforms.deselect(editor)
+              editor.deselect()
             }
 
             processing.current = false
@@ -283,7 +282,7 @@ export const Editable = forwardRef(
             }
 
             if (!domSelection) {
-              return Transforms.deselect(editor)
+              return editor.deselect()
             }
 
             const {anchorNode, focusNode} = domSelection
@@ -306,7 +305,7 @@ export const Editable = forwardRef(
                   !androidInputManager?.hasPendingChanges() &&
                   !androidInputManager?.isFlushing()
                 ) {
-                  Transforms.select(editor, range)
+                  editor.select(range)
                 } else {
                   androidInputManager?.handleUserSelect(range)
                 }
@@ -315,7 +314,7 @@ export const Editable = forwardRef(
 
             // Deselect the editor if the dom selection is not selectable in readonly mode
             if (readOnly && (!anchorNodeSelectable || !focusNodeInEditor)) {
-              Transforms.deselect(editor)
+              editor.deselect()
             }
           }
         }, 100),
@@ -557,7 +556,7 @@ export const Editable = forwardRef(
             suppressThrow: false,
           })
 
-          Transforms.select(editor, slateRange)
+          editor.select(slateRange)
 
           event.preventDefault()
           event.stopImmediatePropagation()
@@ -691,7 +690,7 @@ export const Editable = forwardRef(
                   editor.selection &&
                   Editor.rangeRef(editor, editor.selection)
 
-                Transforms.select(editor, range)
+                editor.select(range)
 
                 if (selectionRef) {
                   editor.userSelection = selectionRef
@@ -904,7 +903,7 @@ export const Editable = forwardRef(
             toRestore &&
             (!editor.selection || !Range.equals(editor.selection, toRestore))
           ) {
-            Transforms.select(editor, toRestore)
+            editor.select(toRestore)
           }
         }
       },
@@ -1310,7 +1309,7 @@ export const Editable = forwardRef(
                         }
 
                         const range = Editor.range(editor, blockPath)
-                        Transforms.select(editor, range)
+                        editor.select(range)
                         return
                       }
 
@@ -1329,7 +1328,7 @@ export const Editable = forwardRef(
                         Path.equals(startVoid[1], endVoid[1])
                       ) {
                         const range = Editor.range(editor, start)
-                        Transforms.select(editor, range)
+                        editor.select(range)
                       }
                     }
                   },
@@ -1495,7 +1494,7 @@ export const Editable = forwardRef(
                             selection.anchor.path,
                           )
                           if (Editor.isVoid(editor, node)) {
-                            Transforms.delete(editor)
+                            editor.delete()
                           }
                         }
                       }
@@ -1542,7 +1541,7 @@ export const Editable = forwardRef(
                       // so that it shows up in the selection's fragment.
                       if (voidMatch) {
                         const range = Editor.range(editor, path)
-                        Transforms.select(editor, range)
+                        editor.select(range)
                       }
 
                       state.isDraggingInternally = true
@@ -1572,7 +1571,7 @@ export const Editable = forwardRef(
                       const range = ReactEditor.findEventRange(editor, event)
                       const data = event.dataTransfer
 
-                      Transforms.select(editor, range)
+                      editor.select(range)
 
                       if (state.isDraggingInternally) {
                         if (
@@ -1580,7 +1579,7 @@ export const Editable = forwardRef(
                           !Range.equals(draggedRange, range) &&
                           !Editor.void(editor, {at: range, voids: true})
                         ) {
-                          Transforms.delete(editor, {
+                          editor.delete({
                             at: draggedRange,
                           })
                         }
@@ -1709,19 +1708,19 @@ export const Editable = forwardRef(
                       // (2017/10/17)
                       if (Hotkeys.isMoveLineBackward(nativeEvent)) {
                         event.preventDefault()
-                        Transforms.move(editor, {unit: 'line', reverse: true})
+                        editor.move({unit: 'line', reverse: true})
                         return
                       }
 
                       if (Hotkeys.isMoveLineForward(nativeEvent)) {
                         event.preventDefault()
-                        Transforms.move(editor, {unit: 'line'})
+                        editor.move({unit: 'line'})
                         return
                       }
 
                       if (Hotkeys.isExtendLineBackward(nativeEvent)) {
                         event.preventDefault()
-                        Transforms.move(editor, {
+                        editor.move({
                           unit: 'line',
                           edge: 'focus',
                           reverse: true,
@@ -1731,7 +1730,7 @@ export const Editable = forwardRef(
 
                       if (Hotkeys.isExtendLineForward(nativeEvent)) {
                         event.preventDefault()
-                        Transforms.move(editor, {unit: 'line', edge: 'focus'})
+                        editor.move({unit: 'line', edge: 'focus'})
                         return
                       }
 
@@ -1744,9 +1743,9 @@ export const Editable = forwardRef(
                         event.preventDefault()
 
                         if (selection && Range.isCollapsed(selection)) {
-                          Transforms.move(editor, {reverse: !isRTL})
+                          editor.move({reverse: !isRTL})
                         } else {
-                          Transforms.collapse(editor, {
+                          editor.collapse({
                             edge: isRTL ? 'end' : 'start',
                           })
                         }
@@ -1758,9 +1757,9 @@ export const Editable = forwardRef(
                         event.preventDefault()
 
                         if (selection && Range.isCollapsed(selection)) {
-                          Transforms.move(editor, {reverse: isRTL})
+                          editor.move({reverse: isRTL})
                         } else {
-                          Transforms.collapse(editor, {
+                          editor.collapse({
                             edge: isRTL ? 'start' : 'end',
                           })
                         }
@@ -1772,10 +1771,10 @@ export const Editable = forwardRef(
                         event.preventDefault()
 
                         if (selection && Range.isExpanded(selection)) {
-                          Transforms.collapse(editor, {edge: 'focus'})
+                          editor.collapse({edge: 'focus'})
                         }
 
-                        Transforms.move(editor, {
+                        editor.move({
                           unit: 'word',
                           reverse: !isRTL,
                         })
@@ -1786,10 +1785,10 @@ export const Editable = forwardRef(
                         event.preventDefault()
 
                         if (selection && Range.isExpanded(selection)) {
-                          Transforms.collapse(editor, {edge: 'focus'})
+                          editor.collapse({edge: 'focus'})
                         }
 
-                        Transforms.move(editor, {
+                        editor.move({
                           unit: 'word',
                           reverse: isRTL,
                         })
