+++ packages/editor/src/slate/editor/insert-break.ts
@@ -1,6 +1,5 @@
 import type {EditorInterface} from '../interfaces/editor'
-import {Transforms} from '../interfaces/transforms'
 
 export const insertBreak: EditorInterface['insertBreak'] = (editor) => {
-  Transforms.splitNodes(editor, {always: true})
+  editor.splitNodes({always: true})
 }
