+++ packages/editor/src/slate/interfaces/transforms/index.ts
@@ -1,14 +1,3 @@
-import {GeneralTransforms} from './general'
-import {NodeTransforms} from './node'
-import {SelectionTransforms} from './selection'
-import {TextTransforms} from './text'
-
-export const Transforms: GeneralTransforms &
-  NodeTransforms &
-  SelectionTransforms &
-  TextTransforms = {
-  ...GeneralTransforms,
-  ...NodeTransforms,
-  ...SelectionTransforms,
-  ...TextTransforms,
-}
+export * from './node'
+export * from './selection'
+export * from './text'
