+++ packages/editor/src/slate/transforms-selection/collapse.ts
@@ -1,5 +1,4 @@
 import {Range} from '../interfaces/range'
-import {Transforms} from '../interfaces/transforms'
 import type {SelectionTransforms} from '../interfaces/transforms/selection'
 
 export const collapse: SelectionTransforms['collapse'] = (
@@ -12,14 +11,14 @@ export const collapse: SelectionTransforms['collapse'] = (
   if (!selection) {
     return
   } else if (edge === 'anchor') {
-    Transforms.select(editor, selection.anchor)
+    editor.select(selection.anchor)
   } else if (edge === 'focus') {
-    Transforms.select(editor, selection.focus)
+    editor.select(selection.focus)
   } else if (edge === 'start') {
     const [start] = Range.edges(selection)
-    Transforms.select(editor, start)
+    editor.select(start)
   } else if (edge === 'end') {
     const [, end] = Range.edges(selection)
-    Transforms.select(editor, end)
+    editor.select(end)
   }
 }
