+++ packages/editor/src/slate/transforms-selection/select.ts
@@ -1,7 +1,6 @@
 import {Editor} from '../interfaces/editor'
 import {Range} from '../interfaces/range'
 import {Scrubber} from '../interfaces/scrubber'
-import {Transforms} from '../interfaces/transforms'
 import type {SelectionTransforms} from '../interfaces/transforms/selection'
 
 export const select: SelectionTransforms['select'] = (editor, target) => {
@@ -9,7 +8,7 @@ export const select: SelectionTransforms['select'] = (editor, target) => {
   target = Editor.range(editor, target)
 
   if (selection) {
-    Transforms.setSelection(editor, target)
+    editor.setSelection(target)
     return
   }
