+++ packages/editor/src/slate/transforms-text/index.ts
@@ -1,2 +1,3 @@
 export * from './delete-text'
 export * from './insert-fragment'
+export * from './insert-text'
